+++ source/Core/Autoload/UnifiedNameSpaceClassMap.php
@@ -3423,4 +3423,16 @@ return [
         'isInterface'      => false,
         'isDeprecated'     => false
     ],
+    'OxidEsales\Eshop\Core\Module\ModuleSmartyPluginDirectories'                                        => [
+        'editionClassName' => \OxidEsales\EshopCommunity\Core\Module\ModuleSmartyPluginDirectories::class,
+        'isAbstract'       => false,
+        'isInterface'      => false,
+        'isDeprecated'     => true
+    ],
+    'OxidEsales\Eshop\Core\Module\ModuleSmartyPluginDirectoryRepository'                                        => [
+        'editionClassName' => \OxidEsales\EshopCommunity\Core\Module\ModuleSmartyPluginDirectoryRepository::class,
+        'isAbstract'       => false,
+        'isInterface'      => false,
+        'isDeprecated'     => true
+    ],
 ];