+++ tests/Integration/Modules/OnlineModuleNotifierTest.php
@@ -76,6 +76,7 @@ class OnlineModuleNotifierTest extends BaseModuleTestCase
         $aModulesInfo[] = array('id' => 'with_events', 'version' => '1.0', 'activeInShop' => array());
         $aModulesInfo[] = array('id' => 'with_everything', 'version' => '1.0', 'activeInShop' => array($sShopUrl));
         $aModulesInfo[] = array('id' => 'with_metadata_v2', 'version' => '1.0', 'activeInShop' => array());
+        $aModulesInfo[] = array('id' => 'with_metadata_v21', 'version' => '1.0', 'activeInShop' => array());
         $aModulesInfo[] = array('id' => 'with_more_metadata_v2', 'version' => '1.0', 'activeInShop' => array());
         $aModulesInfo[] = array('id' => 'without_own_module_namespace', 'version' => '1.0.0', 'activeInShop' => array());